us/24/crud/bad-crud-urls Prepare Delete controller/service for non existant domains/objects urls

+++ grails-app/services/net/kaleidos/plugins/admin/GrailsAdminPluginGenericService.groovy
@@ -49,8 +49,13 @@ class GrailsAdminPluginGenericService {
         return domainClass.count()
     }
 
-    void deleteDomain(Class<?> domainClass, Long objectId){
+    Boolean deleteDomain(Class<?> domainClass, Long objectId){
+        def result = false
         def domainObj = domainClass.get(objectId)
-        domainObj.delete()
+        if (domainObj) {
+            domainObj.delete()
+            result = true
+        }
+        return result
     }
 }
